+++ package.json
@@ -101,11 +101,9 @@
     "stylelint": "~8.0.0",
     "stylelint-config-sass-guidelines": "^3.0.0",
     "stylelint-suitcss": "^1.0.0",
-    "terra-toolkit": "^2.2.0",
+    "terra-toolkit": "^2.3.0",
     "wdio-mocha-framework": "^0.5.11",
-    "wdio-static-server-service": "^1.0.0",
     "wdio-visual-regression-service": "^0.8.0",
-    "wdio-webpack-service": "^1.0.0",
     "webdriverio": "^4.9.8",
     "webpack": "^3.6.0",
     "webpack-dev-server": "2.7.1"
