+++ packages/terra-badge/package.json
@@ -30,7 +30,7 @@
     "props-table": "props-table ./src/Badge.jsx --out-dir ./docs/props-table",
     "test": "npm run test:jest && npm run test:wdio",
     "test:jest": "jest ./tests/jest/* --config ../../jestconfig.json",
-    "test:wdio": "wdio ../../wdio.conf.js --suite badge"
+    "test:wdio": "wdio ../../wdio.conf.js"
   },
   "peerDependencies": {
     "react": "^15.4.2",
